Crate prover

Modules

cli

Structs

BlockProverInput

ProverConfig

Functions

prove

Prove all the blocks in the input, or simulate their execution depending on the selected prover configuration. Return the list of block numbers that are proved and if the proof data is not saved to disk, return the generated block proofs as well.

Type Aliases

BlockProverInputFuture